summer_camp/RAG_first_try/pyproject.toml
[project]
name = "rag-first-try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.16",
    "google-genai>=1.29.0",
    "numpy>=2.3.2",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
